Refactor package structure and package name
Increment A-Packages requires us to divide the project into packages
and subpackages. Also, task list stored in text file sometimes misses
last task added during the session.

Let's,
* divide the project into packages and subpackages
* fix the task history issue
