+++ src/duke/exception/IllegalTaskException.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.exception;
 
 public class IllegalTaskException extends IllegalArgumentException {
     public IllegalTaskException(String s) {
