+++ src/duke/exception/IllegalTokenException.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.exception;
 
 public class IllegalTokenException extends IllegalArgumentException {
     public IllegalTokenException(String s) {
