+++ src/duke/task/Deadline.java
@@ -1,4 +1,6 @@
-package main.java;
+package duke.task;
+
+import duke.task.Task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
